manejo de errores en controllers restore y delete

+++ api/src/controllers/doctors/deleteDoctor.js
@@ -2,6 +2,13 @@ const { DoctorType } = require("../../db");
 
 const deleteDoctor = async ({ id }) => {
   try {
+    const doctor = await DoctorType.findOne({
+      where: { id: id },
+    });
+    if (!doctor) {
+      throw new Error(`No se encontró el doctor con el ID ${id}`);
+    }
+
     await DoctorType.destroy({
       where: {
         id: id,
@@ -10,7 +17,6 @@ const deleteDoctor = async ({ id }) => {
     return "Delete Succesfull";
   } catch (error) {
     throw error;
-    // res.status(400).send(error);
   }
 };
